move test worker listener methods. Fix integration test.
