+++ primary/src/block_synchronizer/tests/block_synchronizer_tests.rs
@@ -21,7 +21,7 @@ use std::{
 };
 use test_utils::{
     certificate, fixture_batch_with_transactions, fixture_header_builder, keys,
-    resolve_name_and_committee, PrimaryToPrimaryMockServer, PrimaryToWorkerMockServer,
+    resolve_name_and_committee, PrimaryToPrimaryMockServer,
 };
 use tokio::{
     sync::mpsc::channel,
@@ -29,6 +29,7 @@ use tokio::{
     time::{sleep, timeout},
 };
 
+use crate::common::worker_listener;
 use tracing::debug;
 use types::{Certificate, CertificateDigest};
 
@@ -810,50 +811,6 @@ where
                 Err(err) => {
                     panic!("Error occurred {err}");
                 }
-            }
-        }
-    })
-}
-
-// TODO: remove this duplication with an associated type
-pub fn worker_listener<T>(
-    num_of_expected_responses: i32,
-    address: multiaddr::Multiaddr,
-) -> JoinHandle<Vec<T>>
-where
-    T: Send + DeserializeOwned + 'static,
-{
-    tokio::spawn(async move {
-        let mut recv = PrimaryToWorkerMockServer::spawn(address);
-        let mut responses = Vec::new();
-
-        loop {
-            match timeout(Duration::from_secs(1), recv.recv()).await {
-                Err(_) => {
-                    // timeout happened - just return whatever has already
-                    return responses;
-                }
-                Ok(Some(message)) => {
-                    match deserialize::<'_, T>(&message.payload) {
-                        Ok(msg) => {
-                            responses.push(msg);
-
-                            // if -1 is given, then we don't count the number of messages
-                            // but we just rely to receive as many as possible until timeout
-                            // happens when waiting for requests.
-                            if num_of_expected_responses != -1
-                                && responses.len() as i32 == num_of_expected_responses
-                            {
-                                return responses;
-                            }
-                        }
-                        Err(err) => {
-                            panic!("Error occurred {err}");
-                        }
-                    }
-                }
-                //  sender closed
-                _ => panic!("Failed to receive network message"),
             }
         }
     })
