+++ primary/src/tests/common.rs
@@ -3,10 +3,14 @@
 use config::WorkerId;
 use crypto::ed25519::Ed25519PublicKey;
 use store::{reopen, rocks, rocks::DBMap, Store};
-use test_utils::{temp_dir, CERTIFICATES_CF, HEADERS_CF, PAYLOAD_CF};
+use test_utils::{temp_dir, PrimaryToWorkerMockServer, CERTIFICATES_CF, HEADERS_CF, PAYLOAD_CF};
 use types::{BatchDigest, Certificate, CertificateDigest, Header, HeaderDigest};
 
 use crate::PayloadToken;
+use bincode::deserialize;
+use serde::de::DeserializeOwned;
+use std::time::Duration;
+use tokio::{task::JoinHandle, time::timeout};
 
 pub fn create_db_stores() -> (
     Store<HeaderDigest, Header<Ed25519PublicKey>>,
@@ -27,4 +31,47 @@ pub fn create_db_stores() -> (
         Store::new(certificate_map),
         Store::new(payload_map),
     )
+}
+
+pub fn worker_listener<T>(
+    num_of_expected_responses: i32,
+    address: multiaddr::Multiaddr,
+) -> JoinHandle<Vec<T>>
+where
+    T: Send + DeserializeOwned + 'static,
+{
+    tokio::spawn(async move {
+        let mut recv = PrimaryToWorkerMockServer::spawn(address);
+        let mut responses = Vec::new();
+
+        loop {
+            match timeout(Duration::from_secs(1), recv.recv()).await {
+                Err(_) => {
+                    // timeout happened - just return whatever has already
+                    return responses;
+                }
+                Ok(Some(message)) => {
+                    match deserialize::<'_, T>(&message.payload) {
+                        Ok(msg) => {
+                            responses.push(msg);
+
+                            // if -1 is given, then we don't count the number of messages
+                            // but we just rely to receive as many as possible until timeout
+                            // happens when waiting for requests.
+                            if num_of_expected_responses != -1
+                                && responses.len() as i32 == num_of_expected_responses
+                            {
+                                return responses;
+                            }
+                        }
+                        Err(err) => {
+                            panic!("Error occurred {err}");
+                        }
+                    }
+                }
+                //  sender closed
+                _ => panic!("Failed to receive network message"),
+            }
+        }
+    })
 }
